pyproject.toml
[tool.poetry]
name = "dutch-fuel-prices"
version = "0.1.0"
description = ""
authors = ["Your Name"]

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
seaborn = "^0.12.2"
ipykernel = "^6.25.2"
cbsodata = "^1.3.4"
statsmodels = "^0.14.0"
mlflow = "^2.7.1"
sktime = "^0.23.0"
pmdarima = "^2.0.3"
darts = "^0.26.0"
prefect = "^2.13.2"
boto3 = "^1.28.54"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
